# docs/api/evalsmith-public.openapi.yaml
openapi: 3.1.0
info:
  title: EvalSmith Public API
  version: 0.1.0
  summary: Gateway-facing API surface for EvalSmith trial and private deployments
  description: >
    First-pass public contract for EvalSmith. This spec focuses on the core
    integration paths external users need for auth, datasets, traces,
    experiments, annotation, and monitoring.
servers:
  - url: http://127.0.0.1:8080
    description: Local gateway
...
paths:
...
  /api/v1/traces:
    get:
      tags: [traces]
      operationId: listTraces
      summary: List traces
      parameters:
        - $ref: "#/components/parameters/ProjectIdHeader"
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: page_size
          in: query
          schema:
            type: integer
            default: 20
        - name: search
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum: [ok, error]
      security:
        - sessionCookie: []
        - bearerApiKey: []
      responses:
        "200":
          description: Trace page
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TraceListEnvelope"
    post:
      tags: [traces]
      operationId: ingestTraces
      summary: Ingest one or more traces
      parameters:
        - $ref: "#/components/parameters/ProjectIdHeader"
      security:
        - sessionCookie: []
        - bearerApiKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BatchIngestRequest"
      responses:
        "200":
          description: Trace batch accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericEnvelope"
  /api/v1/traces/{trace_id}:
    get:
      tags: [traces]
      operationId: getTraceDetail
      summary: Get a single trace with spans
      parameters:
        - $ref: "#/components/parameters/ProjectIdHeader"
        - $ref: "#/components/parameters/TraceIdPath"
      security:
        - sessionCookie: []
        - bearerApiKey: []
      responses:
        "200":
          description: Trace detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericEnvelope"
        "404":
          description: Trace not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /api/v1/traces/batch/dataset:
    post:
      tags: [traces]
      operationId: backfillTracesToDataset
      summary: Create dataset examples from traces
      parameters:
        - $ref: "#/components/parameters/ProjectIdHeader"
      security:
        - sessionCookie: []
        - bearerApiKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TraceBackfillDatasetRequest"
      responses:
        "200":
          description: Dataset backfill result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericEnvelope"
  /api/v1/traces/batch/annotation:
    post:
      tags: [traces]
      operationId: backfillTracesToAnnotation
      summary: Create annotation tasks from traces
      parameters:
        - $ref: "#/components/parameters/ProjectIdHeader"
      security:
        - sessionCookie: []
        - bearerApiKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TraceBackfillAnnotationRequest"
      responses:
        "200":
          description: Annotation backfill result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericEnvelope"
...
  /api/v1/monitoring/rules:
    get:
      tags: [monitoring]
      operationId: listMonitoringRules
      summary: List monitoring rules
      parameters:
        - $ref: "#/components/parameters/ProjectIdHeader"
      security:
        - sessionCookie: []
        - bearerApiKey: []
      responses:
        "200":
          description: Monitoring rules
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericEnvelope"
    post:
      tags: [monitoring]
      operationId: createMonitoringRule
      summary: Create a monitoring rule
      parameters:
        - $ref: "#/components/parameters/ProjectIdHeader"
      security:
        - sessionCookie: []
        - bearerApiKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MonitorRuleCreateRequest"
      responses:
        "200":
          description: Monitoring rule created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericEnvelope"
  /api/v1/monitoring/alerts:
    get:
      tags: [monitoring]
      operationId: listMonitoringAlerts
      summary: List monitoring alerts
      parameters:
        - $ref: "#/components/parameters/ProjectIdHeader"
        - name: status
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      security:
        - sessionCookie: []
        - bearerApiKey: []
      responses:
        "200":
          description: Monitoring alerts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericEnvelope"
  /api/v1/monitoring/runs:
    get:
      tags: [monitoring]
      operationId: listMonitoringRuns
      summary: List recent monitoring runs
      parameters:
        - $ref: "#/components/parameters/ProjectIdHeader"
        - name: rule_id
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      security:
        - sessionCookie: []
        - bearerApiKey: []
      responses:
        "200":
          description: Monitoring runs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericEnvelope"
components:
  securitySchemes:
    bearerApiKey:
      type: http
      scheme: bearer
      __REDACTED_SECRET__: API Key
    sessionCookie:
      type: apiKey
      in: cookie
      name: evalsmith_session
  parameters:
    ProjectIdHeader:
      name: X-Project-ID
      in: header
      required: true
      schema:
        type: string
      description: Active project identifier
...
    TraceIdPath:
      name: trace_id
      in: path
      required: true
      schema:
        type: string
...
  schemas:
    EnvelopeBase:
      type: object
      properties:
        code:
          type: integer
          example: 0
        message:
          type: string
          example: success
        details:
          description: Optional error details
      required: [code, message]
    GenericEnvelope:
      allOf:
        - $ref: "#/components/schemas/EnvelopeBase"
        - type: object
          properties:
            data: {}
    ErrorEnvelope:
      allOf:
        - $ref: "#/components/schemas/EnvelopeBase"
...
    SpanPayload:
      type: object
      properties:
        span_id:
          type: string
        parent_span_id:
          type: string
          nullable: true
        name:
          type: string
        span_type:
          type: string
          enum: [llm, tool, retrieval, chain, agent, custom]
        status:
          type: string
          enum: [ok, error]
        start_time:
          type: string
          format: date-time
        end_time:
          type: string
          format: date-time
        input: {}
        output: {}
        metrics:
          type: object
        metadata:
          type: object
        events:
          type: array
          items:
            type: object
      required: [name, span_type, status, start_time, end_time]
    TracePayload:
      type: object
      properties:
        trace_id:
          type: string
        name:
          type: string
        tags:
          type: array
          items:
            type: string
        metadata:
          type: object
        spans:
          type: array
          items:
            $ref: "#/components/schemas/SpanPayload"
      required: [name, spans]
    BatchIngestRequest:
      type: object
      properties:
        traces:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/TracePayload"
      required: [traces]
    TraceListEnvelope:
      allOf:
        - $ref: "#/components/schemas/EnvelopeBase"
        - type: object
          properties:
            data:
              type: object
              properties:
                traces:
                  type: array
                  items: {}
                total:
                  type: integer
                page:
                  type: integer
                page_size:
                  type: integer
    TraceBackfillDatasetRequest:
      type: object
      properties:
        dataset_id:
          type: string
        trace_ids:
          type: array
          minItems: 1
          items:
            type: string
        split:
          type: string
      required: [dataset_id, trace_ids]
    TraceBackfillAnnotationRequest:
      type: object
      properties:
        trace_ids:
          type: array
          minItems: 1
          items:
            type: string
        mode:
          type: string
      required: [trace_ids]
...
    MonitorRuleCreateRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        status:
          type: string
        sampling_rate:
          type: number
        evaluator_ids:
          type: array
          items:
            type: string
        threshold:
          type: number
        severity:
          type: string
        backfill_dataset_id:
          type: string
          nullable: true
        backfill_split:
          type: string
        auto_annotation:
          type: boolean
        guardrail_config:
          type: object
      required: [name]